requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
python-dotenv>=0.21.0
openai==1.78.1
langchain-openai==0.3.17
tqdm==4.67.1
numpy==1.24.3
langchain_postgres==0.0.14
requests>=2.31.0
flask>=3.0.0